Scheduler time is now 64-bit ns, working on objstore checksum and increasing object length field to 16-bit

+++ include/objstore/objstore.h
@@ -39,7 +39,7 @@ void objstore_init(vmem_t * vmem);
 int objstore_scan(vmem_t * vmem, objstore_scan_callback_f callback, int verbose, void * ctx);
 int objstore_alloc(vmem_t * vmem, int length, int verbose);
 
-void objstore_write_obj(vmem_t * vmem, int offset, uint8_t type, uint8_t length, void * data);
+void objstore_write_obj(vmem_t * vmem, int offset, uint8_t type, uint16_t length, void * data);
 int objstore_rm_obj(vmem_t * vmem, int offset, int verbose);
 
 int objstore_read_obj_type(vmem_t * vmem, int offset);
